upgrade to fit windows

+++ install.sh
@@ -1,38 +1,46 @@
 #!/bin/bash
 
 #exit immediatly
-#set -e
+set +e
 #no emit unset var
 set -u
 
-CURRENT_DIR=$(cd `dirname $0`;pwd)
-source "$CURRENT_DIR/my.bashrc"
-#check path exsists
+CURRENT_PATH=$(cd `dirname $0`;pwd)
+source "$CURRENT_PATH/my.bashrc"
+#Check path exsists
+echo "------- Checking releated paths are created---------"
+echo "checking \$MY_PATH:$MY_PATH"
 if [ ! -d $MY_PATH ];then
+	echo "\$MY_PATH not exsisted, try to create"
 	mkdir -p $MY_PATH
 fi
+echo "checking \$GOPATH:$GOPATH"
 if [ ! -d $GOPATH ];then
+	echo "\$GOPATH not exsisted, try to create"
 	mkdir -p $GOPATH
 fi
-if [ $CURRENT_DIR != $MY_BASHRC_PATH ];then
-	mv $CURRENT_DIR $MY_PATH
+
+echo "checking \$Current_PATH and \$MY_BASHRC_PATH"
+
+if [ $CURRENT_PATH != $MY_BASHRC_PATH ];then
+	echo "  CURRENT_PATH($CURRENT_PATH) != MY_BASHRC_PATH($MY_BASHRC_PATH), try to move it" 
+	mv $CURRENT_PATH $MY_BASHRC_PATH
 fi
-#echo "Current Path: $CURRENT_DIR"
-#echo "MY_BASHRC_PATH: $MY_BASHRC_PATH"
-#echo "pwd: `pwd`"
-add="
-#Source my bashrc
+
+#Check my bashrc is installed
+echo "------- Checking mybashrc is installed---------"
+
+add="#Source my bashrc
 export MY_BASHRC_FILE=$MY_BASHRC_FILE
 if [ -f \$MY_BASHRC_FILE ];then
 	source \$MY_BASHRC_FILE
-fi
-"
+fi"
+
 check="grep \"Source my bashrc\" $SYSTEM_BASHRC_FILE"
 check_result=`eval $check`
 echo "check cmd: $check"
 echo "check result: $check_result"
 
-#:<<!
 if [ "$check_result" == "" ];then
 	echo "You have not installed mybashrc"
 	echo "$add" >> $SYSTEM_BASHRC_FILE
@@ -41,13 +49,22 @@ else
 	echo "You have installed mybashrc"
 fi
 echo "use **source $SYSTEM_BASHRC_FILE** to apply, then use **bashrc** to modif\n"
-#!
+
+#Check OS
+echo "------- Checking OS---------"
+cmd.exe /c ver > /dev/null
+if [ $? -eq 0 ];then
+	OS="Windows"
+else
+	OS="Linux"
+fi
+echo "OS Type: $OS"
 
 #Install Golang
 echo "--------Tring to install Golang---------"
-if [ ! -d $GOROOT ];then
+if [ ! -d $GOROOT ] && [ $OS == "Linux"];then
 	echo "You have not installed Golang"
-	read -p "Would you like to upgrade golang to $GOVERSION (y/no):" res
+	read -p "Would you like to install golang $GOVERSION (y/no):" res
 	if [ $res == "y" ];then
 		filename="go${GOVERSION}.linux-amd64.tar.gz"
 		filepath="/root/$filename"
@@ -65,33 +82,36 @@ fi
 #Install Git
 echo "--------Tring to install Git"
 echo "Your git version: `git --version`"
-read -p "Would you like to upgrade git to $GITVERSION (y/no):" res
-if [ $res == "y" ];then
-	installdir="/tmp/sel/git"
-	filename="git-${GITVERSION}.tar.xz"
-	filepath="$installdir/$filename"
-	link="https://www.kernel.org/pub/software/scm/git/${filename}"
-	untardir="$installdir/untar"
-	if [ ! -d $installdir ];then
-		mkdir -p $installdir
+if [ $OS == "Linux" ];then 
+	read -p "Would you like to upgrade git to $GITVERSION (y/no):" res
+	if [ $res == "y" ];then
+		installdir="/tmp/sel/git"
+		filename="git-${GITVERSION}.tar.xz"
+		filepath="$installdir/$filename"
+		link="https://www.kernel.org/pub/software/scm/git/${filename}"
+		untardir="$installdir/untar"
+		if [ ! -d $installdir ];then
+			mkdir -p $installdir
+		fi
+		if [ ! -f $filepath ];then
+			wget -O $filepath $link
+		fi
+		if [ ! -d $untardir ];then
+			mkdir -p $untardir
+			tar -C $untardir -xf $filepath
+		fi
+		cd $untardir/git-${GITVERSION}
+		#yum install -y curl-devel expat-devel gettext-devel openssl-devel zlib-devel asciidoc
+		#yum install -y gcc perl-ExtUtils-MakeMaker
+		yum remove git -y
+		rm -rf $GITROOT/*
+		make prefix=$GITROOT all
+		make prefix=$GITROOT install
+		rm -rf $installdir
 	fi
-	if [ ! -f $filepath ];then
-		wget -O $filepath $link
-	fi
-	if [ ! -d $untardir ];then
-		mkdir -p $untardir
-		tar -C $untardir -xf $filepath
-	fi
-	cd $untardir/git-${GITVERSION}
-	#yum install -y curl-devel expat-devel gettext-devel openssl-devel zlib-devel asciidoc
-	#yum install -y gcc perl-ExtUtils-MakeMaker
-	yum remove git -y
-	rm -rf $GITROOT/*
-	make prefix=$GITROOT all
-	make prefix=$GITROOT install
-	rm -rf $installdir
 fi
 #Install KubeEdge-src
+<<'!'
 echo "--------Tring to install kubeedge src--------"
 if [ ! -d $KUBEEDGE_PATH ];then
 	echo "You have not installed kubeedge src"
@@ -104,4 +124,5 @@ if [ ! -d $KUBEEDGE_PATH ];then
 else
 	echo "You have installed kubeedge src"
 fi
+!
